+++ src/Entity/Organisasi/Kantor.php
@@ -45,7 +45,8 @@ use Symfony\Component\Validator\Constraints as Assert;
         'id',
         'legacy_kode',
         'legacy_kode_kpp',
-        'legacy_kode_kanwil'
+        'legacy_kode_kanwil',
+        'ministry_office_code',
     ],
     name: 'idx_kantor_legacy'
 )]
@@ -175,6 +176,7 @@ use Symfony\Component\Validator\Constraints as Assert;
         'kecamatanName' => 'iexact',
         'kelurahan' => 'iexact',
         'kelurahanName' => 'iexact',
+        'ministryOfficeCode' => 'exact',
     ]
 )]
 #[ApiFilter(
@@ -539,7 +541,7 @@ class Kantor
             'kantor:write'
         ]
     )]
-    private ?string $KabupatenKotaName = null;
+    private ?string $kabupatenKotaName = null;
 
     #[ORM\Column(
         type: Types::STRING,
@@ -552,7 +554,7 @@ class Kantor
             'kantor:write'
         ]
     )]
-    private ?string $KecamatanName = null;
+    private ?string $kecamatanName = null;
 
     #[ORM\Column(
         type: Types::STRING,
@@ -565,7 +567,20 @@ class Kantor
             'kantor:write'
         ]
     )]
-    private ?string $KelurahanName = null;
+    private ?string $kelurahanName = null;
+
+    #[ORM\Column(
+        type: Types::STRING,
+        length: 10,
+        nullable: true
+    )]
+    #[Groups(
+        groups: [
+            'kantor:read',
+            'kantor:write'
+        ]
+    )]
+    private ?string $ministryOfficeCode = null;
 
     public function __construct()
     {
@@ -998,36 +1013,48 @@ class Kantor
 
     public function getKabupatenKotaName(): ?string
     {
-        return $this->KabupatenKotaName;
+        return $this->kabupatenKotaName;
     }
 
-    public function setKabupatenKotaName(?string $KabupatenKotaName): self
+    public function setKabupatenKotaName(?string $kabupatenKotaName): self
     {
-        $this->KabupatenKotaName = $KabupatenKotaName;
+        $this->kabupatenKotaName = $kabupatenKotaName;
 
         return $this;
     }
 
     public function getKecamatanName(): ?string
     {
-        return $this->KecamatanName;
+        return $this->kecamatanName;
     }
 
-    public function setKecamatanName(?string $KecamatanName): self
+    public function setKecamatanName(?string $kecamatanName): self
     {
-        $this->KecamatanName = $KecamatanName;
+        $this->kecamatanName = $kecamatanName;
 
         return $this;
     }
 
     public function getKelurahanName(): ?string
     {
-        return $this->KelurahanName;
+        return $this->kelurahanName;
     }
 
-    public function setKelurahanName(?string $KelurahanName): self
+    public function setKelurahanName(?string $kelurahanName): self
     {
-        $this->KelurahanName = $KelurahanName;
+        $this->kelurahanName = $kelurahanName;
+
+        return $this;
+    }
+
+    public function getMinistryOfficeCode(): ?string
+    {
+        return $this->ministryOfficeCode;
+    }
+
+    public function setMinistryOfficeCode(?string $ministryOfficeCode): self
+    {
+        $this->ministryOfficeCode = $ministryOfficeCode;
 
         return $this;
     }